qemu_v8: use tee-raw.bin in sp_layout.json
When configured for Hafnium as SPMC at S-EL2 OP-TEE is packaged with
help from the file sp_layout.json. The JSON file points out the OP-TEE
binary to load. tee-pager_v2.bin was used until this patch, but the name
of that file is unintuitive and confusing. So fix this by using the
preferred file tee-raw.bin.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>