qemu_v8: xen-create-image: compute root FS size
Since the Buildroot upgrade [1], it looks like the Xen root FS is a bit
bigger resulting in:

 __populate_fs: Could not allocate block in ext2 filesystem while writing file "rootfs.cpio.gz"

Therefore compute the image size from size of the directory we copy into
the image. Some overhead has to be taken into account, that is why we
multiply by 1.2. Currently this leaves about 9M of free space (2412
times 4K), and SZ is 111M.

  $ dumpe2fs -h ../out/bin/xen.ext4 | grep -E '(Free blocks|Block size)'
  Free blocks:              2412
  Block size:               4096

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Link: https://github.com/OP-TEE/manifest/commit/aab39982ed1219b33ddae75c85953036de085eb2 [1]
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>